+++ app-angular/src/app/interval/fromMsToMm.ts
